refactor(ItineraryExistenceChecker): Merge class into ItineraryUtils.
